reversedropshipping/requirements.txt
yt-dlp>=2023.7.6
moviepy>=1.0.3
requests>=2.31.0
pillow>=10.0.0
